+++ src/CorsService.php
@@ -12,55 +12,123 @@
 
 namespace Fruitcake\Cors;
 
+use Fruitcake\Cors\Exceptions\InvalidOptionException;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
+/**
+ * @phpstan-type CorsInputOptions array{
+ *  'allowedOrigins'?: string[],
+ *  'allowedOriginsPatterns'?: string[],
+ *  'supportsCredentials'?: bool,
+ *  'allowedHeaders'?: string[],
+ *  'allowedMethods'?: string[],
+ *  'exposedHeaders'?: string[]|false,
+ *  'maxAge'?: int|bool|null,
+ *  'allowed_origins'?: string[],
+ *  'allowed_origins_patterns'?: string[],
+ *  'supports_credentials'?: bool,
+ *  'allowed_headers'?: string[],
+ *  'allowed_methods'?: string[],
+ *  'exposed_headers'?: string[]|false,
+ *  'max_age'?: int|bool|null
+ * }
+ *
+ */
 class CorsService
 {
-    private $options;
+    /** @var string[]  */
+    private array $allowedOrigins = [];
+    /** @var string[] */
+    private array $allowedOriginsPatterns = [];
+    /** @var string[] */
+    private array $allowedMethods = [];
+    /** @var string[] */
+    private array $allowedHeaders = [];
+    /** @var string[] */
+    private array $exposedHeaders = [];
+    private bool $supportsCredentials = false;
+    private ?int $maxAge = 0;
 
+    private bool $allowAllOrigins = false;
+    private bool $allowAllMethods = false;
+    private bool $allowAllHeaders = false;
+
+    /**
+     * @param CorsInputOptions $options
+     */
     public function __construct(array $options = [])
     {
-        $this->options = $this->normalizeOptions($options);
-    }
-
-    private function normalizeOptions(array $options = []): array
-    {
-        $options += [
-            'allowedOrigins' => [],
-            'allowedOriginsPatterns' => [],
-            'supportsCredentials' => false,
-            'allowedHeaders' => [],
-            'exposedHeaders' => [],
-            'allowedMethods' => [],
-            'maxAge' => 0,
-        ];
-
-        // normalize array('*') to true
-        if (in_array('*', $options['allowedOrigins'])) {
-            $options['allowedOrigins'] = true;
+        if ($options) {
+            $this->setOptions($options);
         }
-        if (in_array('*', $options['allowedHeaders'])) {
-            $options['allowedHeaders'] = true;
-        } else {
-            $options['allowedHeaders'] = array_map('strtolower', $options['allowedHeaders']);
-        }
-
-        if (in_array('*', $options['allowedMethods'])) {
-            $options['allowedMethods'] = true;
-        } else {
-            $options['allowedMethods'] = array_map('strtoupper', $options['allowedMethods']);
-        }
-
-        return $options;
     }
 
     /**
-     * @deprecated use isOriginAllowed
+     * @param CorsInputOptions $options
      */
-    public function isActualRequestAllowed(Request $request): bool
+    public function setOptions(array $options): void
     {
-        return $this->isOriginAllowed($request);
+        $this->allowedOrigins = $options['allowedOrigins'] ?? $options['allowed_origins'] ?? $this->allowedOrigins;
+        $this->allowedOriginsPatterns =
+            $options['allowedOriginsPatterns'] ?? $options['allowed_origins_patterns'] ?? $this->allowedOriginsPatterns;
+        $this->allowedMethods = $options['allowedMethods'] ?? $options['allowed_methods'] ?? $this->allowedMethods;
+        $this->allowedHeaders = $options['allowedHeaders'] ?? $options['allowed_headers'] ?? $this->allowedHeaders;
+        $this->supportsCredentials =
+            $options['supportsCredentials'] ?? $options['supports_credentials'] ?? $this->supportsCredentials;
+
+        $maxAge = $this->maxAge;
+        if (array_key_exists('maxAge', $options)) {
+            $maxAge = $options['maxAge'];
+        } elseif (array_key_exists('max_age', $options)) {
+            $maxAge = $options['max_age'];
+        }
+        $this->maxAge = $maxAge === null ? null : (int)$maxAge;
+
+        $exposedHeaders = $options['exposedHeaders'] ?? $options['exposed_headers'] ?? $this->exposedHeaders;
+        $this->exposedHeaders = $exposedHeaders === false ? [] : $exposedHeaders;
+
+        $this->normalizeOptions();
+    }
+
+    private function normalizeOptions(): void
+    {
+        // Normalize case
+        $this->allowedHeaders = array_map('strtolower', $this->allowedHeaders);
+        $this->allowedMethods = array_map('strtoupper', $this->allowedMethods);
+
+        // Normalize ['*'] to true
+        $this->allowAllOrigins = in_array('*', $this->allowedOrigins);
+        $this->allowAllHeaders = in_array('*', $this->allowedHeaders);
+        $this->allowAllMethods = in_array('*', $this->allowedMethods);
+
+        // Transform wildcard pattern
+        if (!$this->allowAllOrigins) {
+            foreach ($this->allowedOrigins as $origin) {
+                if (strpos($origin, '*') !== false) {
+                    $this->allowedOriginsPatterns[] = $this->convertWildcardToPattern($origin);
+                }
+            }
+        }
+    }
+
+    /**
+     * Create a pattern for a wildcard, based on Str::is() from Laravel
+     *
+     * @see https://github.com/laravel/framework/blob/5.5/src/Illuminate/Support/Str.php
+     * @param string $pattern
+     * @return string
+     */
+    private function convertWildcardToPattern($pattern)
+    {
+        $pattern = preg_quote($pattern, '#');
+
+        // Asterisks are translated into zero-or-more regular expression wildcards
+        // to make it convenient to check if the strings starts with the given
+        // pattern such as "*.example.com", making any string check convenient.
+        $pattern = str_replace('\*', '.*', $pattern);
+
+        return '#^' . $pattern . '\z#u';
     }
 
     public function isCorsRequest(Request $request): bool
@@ -101,21 +169,17 @@ class CorsService
 
     public function isOriginAllowed(Request $request): bool
     {
-        if ($this->options['allowedOrigins'] === true) {
+        if ($this->allowAllOrigins === true) {
             return true;
         }
 
-        if (!$request->headers->has('Origin')) {
-            return false;
-        }
+        $origin = (string) $request->headers->get('Origin');
 
-        $origin = $request->headers->get('Origin');
-
-        if (in_array($origin, $this->options['allowedOrigins'])) {
+        if (in_array($origin, $this->allowedOrigins)) {
             return true;
         }
 
-        foreach ($this->options['allowedOriginsPatterns'] as $pattern) {
+        foreach ($this->allowedOriginsPatterns as $pattern) {
             if (preg_match($pattern, $origin)) {
                 return true;
             }
@@ -137,18 +201,18 @@ class CorsService
         return $response;
     }
 
-    private function configureAllowedOrigin(Response $response, Request $request)
+    private function configureAllowedOrigin(Response $response, Request $request): void
     {
-        if ($this->options['allowedOrigins'] === true && !$this->options['supportsCredentials']) {
+        if ($this->allowAllOrigins === true && !$this->supportsCredentials) {
             // Safe+cacheable, allow everything
             $response->headers->set('Access-Control-Allow-Origin', '*');
         } elseif ($this->isSingleOriginAllowed()) {
             // Single origins can be safely set
-            $response->headers->set('Access-Control-Allow-Origin', array_values($this->options['allowedOrigins'])[0]);
+            $response->headers->set('Access-Control-Allow-Origin', array_values($this->allowedOrigins)[0]);
         } else {
             // For dynamic headers, set the requested Origin header when set and allowed
             if ($this->isCorsRequest($request) && $this->isOriginAllowed($request)) {
-                $response->headers->set('Access-Control-Allow-Origin', $request->headers->get('Origin'));
+                $response->headers->set('Access-Control-Allow-Origin', (string) $request->headers->get('Origin'));
             }
 
             $this->varyHeader($response, 'Origin');
@@ -157,70 +221,65 @@ class CorsService
 
     private function isSingleOriginAllowed(): bool
     {
-        if ($this->options['allowedOrigins'] === true || !empty($this->options['allowedOriginsPatterns'])) {
+        if ($this->allowAllOrigins === true || count($this->allowedOriginsPatterns) > 0) {
             return false;
         }
 
-        return count($this->options['allowedOrigins']) === 1;
+        return count($this->allowedOrigins) === 1;
     }
 
-    private function configureAllowedMethods(Response $response, Request $request)
+    private function configureAllowedMethods(Response $response, Request $request): void
     {
-        if ($this->options['allowedMethods'] === true) {
-            $allowMethods = strtoupper($request->headers->get('Access-Control-Request-Method'));
+        if ($this->allowAllMethods === true) {
+            $allowMethods = strtoupper((string) $request->headers->get('Access-Control-Request-Method'));
             $this->varyHeader($response, 'Access-Control-Request-Method');
         } else {
-            $allowMethods = implode(', ', $this->options['allowedMethods']);
+            $allowMethods = implode(', ', $this->allowedMethods);
         }
 
         $response->headers->set('Access-Control-Allow-Methods', $allowMethods);
     }
 
-    private function configureAllowedHeaders(Response $response, Request $request)
+    private function configureAllowedHeaders(Response $response, Request $request): void
     {
-        if ($this->options['allowedHeaders'] === true) {
-            $allowHeaders = $request->headers->get('Access-Control-Request-Headers');
+        if ($this->allowAllHeaders === true) {
+            $allowHeaders = (string) $request->headers->get('Access-Control-Request-Headers');
             $this->varyHeader($response, 'Access-Control-Request-Headers');
         } else {
-            $allowHeaders = implode(', ', $this->options['allowedHeaders']);
+            $allowHeaders = implode(', ', $this->allowedHeaders);
         }
         $response->headers->set('Access-Control-Allow-Headers', $allowHeaders);
     }
 
-    private function configureAllowCredentials(Response $response, Request $request)
+    private function configureAllowCredentials(Response $response, Request $request): void
     {
-        if ($this->options['supportsCredentials']) {
+        if ($this->supportsCredentials) {
             $response->headers->set('Access-Control-Allow-Credentials', 'true');
         }
     }
 
-    private function configureExposedHeaders(Response $response, Request $request)
+    private function configureExposedHeaders(Response $response, Request $request): void
     {
-        if ($this->options['exposedHeaders']) {
-            $response->headers->set('Access-Control-Expose-Headers', implode(', ', $this->options['exposedHeaders']));
+        if ($this->exposedHeaders) {
+            $response->headers->set('Access-Control-Expose-Headers', implode(', ', $this->exposedHeaders));
         }
     }
 
-    private function configureMaxAge(Response $response, Request $request)
+    private function configureMaxAge(Response $response, Request $request): void
     {
-        if ($this->options['maxAge'] !== null) {
-            $response->headers->set('Access-Control-Max-Age', (int) $this->options['maxAge']);
+        if ($this->maxAge !== null) {
+            $response->headers->set('Access-Control-Max-Age', (string) $this->maxAge);
         }
     }
 
-    public function varyHeader(Response $response, $header): Response
+    public function varyHeader(Response $response, string $header): Response
     {
         if (!$response->headers->has('Vary')) {
             $response->headers->set('Vary', $header);
-        } elseif (!in_array($header, explode(', ', $response->headers->get('Vary')))) {
-            $response->headers->set('Vary', $response->headers->get('Vary') . ', ' . $header);
+        } elseif (!in_array($header, explode(', ', (string) $response->headers->get('Vary')))) {
+            $response->headers->set('Vary', ((string) $response->headers->get('Vary')) . ', ' . $header);
         }
 
         return $response;
-    }
-
-    private function isSameHost(Request $request): bool
-    {
-        return $request->headers->get('Origin') === $request->getSchemeAndHttpHost();
     }
 }