+++ tests/CorsTest.php
@@ -17,12 +17,15 @@ use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 
+/**
+ * @phpstan-import-type CorsInputOptions from CorsService
+ */
 class CorsTest extends TestCase
 {
     /**
      * @test
      */
-    public function it_does_modify_on_a_request_without_origin()
+    public function itDoesModifyOnARequestWithoutOrigin(): void
     {
         $app = $this->createStackedApp();
 
@@ -34,7 +37,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_does_modify_on_a_request_with_same_origin()
+    public function itDoesModifyOnARequestWithSameOrigin(): void
     {
         $app = $this->createStackedApp(array('allowedOrigins' => array('*')));
         $unmodifiedResponse = new Response();
@@ -50,7 +53,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_returns_allow_origin_header_on_valid_actual_request()
+    public function itReturnsAllowOriginHeaderOnValidActualRequest(): void
     {
         $app      = $this->createStackedApp();
         $request  = $this->createValidActualRequest();
@@ -64,7 +67,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_returns_allow_origin_header_on_allow_all_origin_request()
+    public function itReturnsAllowOriginHeaderOnAllowAllOriginRequest(): void
     {
         $app      = $this->createStackedApp(array('allowedOrigins' => array('*')));
         $request  = new Request();
@@ -80,7 +83,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_returns_allow_headers_header_on_allow_all_headers_request()
+    public function itReturnsAllowHeadersHeaderOnAllowAllHeadersRequest(): void
     {
         $app     = $this->createStackedApp(array('allowedHeaders' => array('*')));
         $request = $this->createValidPreflightRequest();
@@ -96,7 +99,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_returns_allow_headers_header_on_allow_all_headers_request_credentials()
+    public function itReturnsAllowHeadersHeaderOnAllowAllHeadersRequestCredentials(): void
     {
         $app      = $this->createStackedApp(array('allowedHeaders' => array('*'), 'supportsCredentials' => true));
         $request = $this->createValidPreflightRequest();
@@ -112,7 +115,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_sets_allow_credentials_header_when_flag_is_set_on_valid_actual_request()
+    public function itSetsAllowCredentialsHeaderWhenFlagIsSetOnValidActualRequest(): void
     {
         $app     = $this->createStackedApp(array('supportsCredentials' => true));
         $request = $this->createValidActualRequest();
@@ -126,7 +129,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_does_not_set_allow_credentials_header_when_flag_is_not_set_on_valid_actual_request()
+    public function itDoesNotSetAllowCredentialsHeaderWhenFlagIsNotSetOnValidActualRequest(): void
     {
         $app     = $this->createStackedApp();
         $request = $this->createValidActualRequest();
@@ -139,7 +142,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_sets_exposed_headers_when_configured_on_actual_request()
+    public function itSetsExposedHeadersWhenConfiguredOnActualRequest(): void
     {
         $app     = $this->createStackedApp(array('exposedHeaders' => array('x-exposed-header', 'x-another-exposed-header')));
         $request = $this->createValidActualRequest();
@@ -153,7 +156,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_adds_a_vary_header_when_wildcard_and_supports_credentials()
+    public function itAddsAVaryHeaderWhenWildcardAndSupportsCredentials(): void
     {
         $app      = $this->createStackedApp(array(
             'allowedOrigins' => ['*'],
@@ -170,7 +173,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_adds_multiple_vary_header_when_wildcard_and_supports_credentials()
+    public function itAddsMultipleVaryHeaderWhenWildcardAndSupportsCredentials(): void
     {
         $app = $this->createStackedApp(array(
             'allowedOrigins' => ['*'],
@@ -188,7 +191,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_adds_a_vary_header_when_has_origin_patterns()
+    public function itAddsAVaryHeaderWhenHasOriginPatterns(): void
     {
         $app      = $this->createStackedApp(array(
             'allowedOriginsPatterns' => array('/l(o|0)calh(o|0)st/')
@@ -204,7 +207,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_doesnt_add_a_vary_header_when_wilcard_origins()
+    public function itDoesntAddAVaryHeaderWhenWilcardOrigins(): void
     {
         $app      = $this->createStackedApp(array(
             'allowedOrigins' => array('*', 'http://localhost')
@@ -219,7 +222,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_doesnt_add_a_vary_header_when_simple_origins()
+    public function itDoesntAddAVaryHeaderWhenSimpleOrigins(): void
     {
         $app = $this->createStackedApp(array(
             'allowedOrigins' => array('http://localhost')
@@ -235,7 +238,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_adds_a_vary_header_when_multiple_origins()
+    public function itAddsAVaryHeaderWhenMultipleOrigins(): void
     {
         $app = $this->createStackedApp(array(
            'allowedOrigins' => array('http://localhost', 'http://example.com')
@@ -252,7 +255,7 @@ class CorsTest extends TestCase
      * @test
      * @see http://www.w3.org/TR/cors/index.html#resource-implementation
      */
-    public function it_appends_an_existing_vary_header()
+    public function itAppendsAnExistingVaryHeader(): void
     {
         $app      = $this->createStackedApp(
             array(
@@ -274,7 +277,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_returns_access_control_headers_on_cors_request()
+    public function itReturnsAccessControlHeadersOnCorsRequest(): void
     {
         $app      = $this->createStackedApp();
         $request  = new Request();
@@ -289,7 +292,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_returns_access_control_headers_on_cors_request_with_pattern_origin()
+    public function itReturnsAccessControlHeadersOnCorsRequestWithPatternOrigin(): void
     {
         $app = $this->createStackedApp(array(
           'allowedOrigins' => array(),
@@ -308,7 +311,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_adds_vary_headers_on_preflight_non_preflight_options()
+    public function itAddsVaryHeadersOnPreflightNonPreflightOptions(): void
     {
         $app      = $this->createStackedApp();
         $request  = new Request();
@@ -322,7 +325,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_returns_access_control_headers_on_valid_preflight_request()
+    public function itReturnsAccessControlHeadersOnValidPreflightRequest(): void
     {
         $app     = $this->createStackedApp();
         $request = $this->createValidPreflightRequest();
@@ -337,7 +340,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_does_not_allow_request_with_origin_not_allowed()
+    public function itDoesNotAllowRequestWithOriginNotAllowed(): void
     {
         $passedOptions = array(
           'allowedOrigins' => array('http://notlocalhost'),
@@ -354,7 +357,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_does_not_modify_request_with_pattern_origin_not_allowed()
+    public function itDoesNotModifyRequestWithPatternOriginNotAllowed(): void
     {
         $passedOptions = array(
             'allowedOrigins' => array(),
@@ -372,7 +375,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_allow_methods_on_valid_preflight_request()
+    public function itAllowMethodsOnValidPreflightRequest(): void
     {
         $app     = $this->createStackedApp(array('allowedMethods' => array('get', 'put')));
         $request = $this->createValidPreflightRequest();
@@ -387,7 +390,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_returns_valid_preflight_request_with_allow_methods_all()
+    public function itReturnsValidPreflightRequestWithAllowMethodsAll(): void
     {
         $app     = $this->createStackedApp(array('allowedMethods' => array('*')));
         $request = $this->createValidPreflightRequest();
@@ -398,13 +401,12 @@ class CorsTest extends TestCase
         // it will return the Access-Control-Request-Method pass in the request
         $this->assertEquals('GET', $response->headers->get('Access-Control-Allow-Methods'));
         $this->assertEquals('Access-Control-Request-Method', $response->headers->get('Vary'));
-
     }
 
     /**
      * @test
      */
-    public function it_returns_valid_preflight_request_with_allow_methods_all_credentials()
+    public function itReturnsValidPreflightRequestWithAllowMethodsAllCredentials(): void
     {
         $app     = $this->createStackedApp(array('allowedMethods' => array('*'), 'supportsCredentials' => true));
         $request = $this->createValidPreflightRequest();
@@ -421,7 +423,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_returns_ok_on_valid_preflight_request_with_requested_headers_allowed()
+    public function itReturnsOkOnValidPreflightRequestWithRequestedHeadersAllowed(): void
     {
         $app            = $this->createStackedApp();
         $requestHeaders = 'X-Allowed-Header, x-other-allowed-header';
@@ -440,7 +442,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_sets_allow_credentials_header_when_flag_is_set_on_valid_preflight_request()
+    public function itSetsAllowCredentialsHeaderWhenFlagIsSetOnValidPreflightRequest(): void
     {
         $app     = $this->createStackedApp(array('supportsCredentials' => true));
         $request = $this->createValidPreflightRequest();
@@ -454,7 +456,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_does_not_set_allow_credentials_header_when_flag_is_not_set_on_valid_preflight_request()
+    public function itDoesNotSetAllowCredentialsHeaderWhenFlagIsNotSetOnValidPreflightRequest(): void
     {
         $app     = $this->createStackedApp();
         $request = $this->createValidPreflightRequest();
@@ -467,7 +469,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_sets_max_age_when_set()
+    public function itSetsMaxAgeWhenSet(): void
     {
         $app     = $this->createStackedApp(array('maxAge' => 42));
         $request = $this->createValidPreflightRequest();
@@ -481,7 +483,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_sets_max_age_when_zero()
+    public function itSetsMaxAgeWhenZero(): void
     {
         $app     = $this->createStackedApp(array('maxAge' => 0));
         $request = $this->createValidPreflightRequest();
@@ -495,7 +497,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_doesnt_set_max_age_when_false()
+    public function itDoesntSetMaxAgeWhenFalse(): void
     {
         $app     = $this->createStackedApp(array('maxAge' => null));
         $request = $this->createValidPreflightRequest();
@@ -508,7 +510,7 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_skips_empty_access_control_request_header()
+    public function itSkipsEmptyAccessControlRequestHeader(): void
     {
         $app     = $this->createStackedApp();
         $request = $this->createValidPreflightRequest();
@@ -521,19 +523,19 @@ class CorsTest extends TestCase
     /**
      * @test
      */
-    public function it_doesnt_set_access_control_allow_origin_without_origin()
+    public function itDoesntSetAccessControlAllowOriginWithoutOrigin(): void
     {
         $app     = $this->createStackedApp([
             'allowedOrigins'      => ['*'],
             'supportsCredentials' => true,
         ]);
 
-        $response = $app->handle(new Request);
+        $response = $app->handle(new Request());
 
         $this->assertFalse($response->headers->has('Access-Control-Allow-Origin'));
     }
 
-    private function createValidActualRequest()
+    private function createValidActualRequest(): Request
     {
         $request  = new Request();
         $request->headers->set('Origin', 'http://localhost');
@@ -541,7 +543,7 @@ class CorsTest extends TestCase
         return $request;
     }
 
-    private function createValidPreflightRequest()
+    private function createValidPreflightRequest(): Request
     {
         $request  = new Request();
         $request->headers->set('Origin', 'http://localhost');
@@ -551,19 +553,17 @@ class CorsTest extends TestCase
         return $request;
     }
 
-    private function createStackedApp(array $options = array(), array $responseHeaders = array())
+    /**
+     * @param CorsInputOptions $options
+     * @param string[] $responseHeaders
+     * @return MockApp
+     */
+    private function createStackedApp(array $options = array(), array $responseHeaders = array()): MockApp
     {
-        $passedOptions = array_merge(array(
-                'allowedHeaders'      => array('x-allowed-header', 'x-other-allowed-header'),
-                'allowedMethods'      => array('delete', 'get', 'post', 'put'),
-                'allowedOrigins'      => array('http://localhost'),
-                'exposedHeaders'      => false,
-                'maxAge'              => false,
-                'supportsCredentials' => false,
-            ),
-            $options
-        );
+        $options['allowedHeaders'] = $options['allowedHeaders'] ?? ['x-allowed-header', 'x-other-allowed-header'];
+        $options['allowedMethods'] = $options['allowedMethods'] ?? ['delete', 'get', 'post', 'put'];
+        $options['allowedOrigins'] = $options['allowedOrigins'] ?? ['http://localhost'];
 
-        return new MockApp($responseHeaders, $passedOptions);
+        return new MockApp($responseHeaders, $options);
     }
 }